requirements.txt
torch==2.2.0
torchvision==0.17.0
flask==3.0.2
pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
python-dotenv==1.0.1 